requirements.txt
streamlit>=1.32.0
plotly>=5.20.0
yfinance>=0.2.37
pandas>=2.2.0
numpy>=1.26.0
xgboost>=2.0.0
scikit-learn>=1.4.0
ta>=0.11.0
requests>=2.31.0
fredapi>=0.5.1
scipy>=1.12.0
pyarrow>=15.0.0
python-dotenv>=1.0.0
lightgbm>=4.3.0
joblib>=1.3.0
